+++ src/core/register.js
@@ -1,7 +1,10 @@
 
 import {
   MODULE_DEFAULT, MODULE_GLOBAL, ERR, CHANGE_BY_SELF, SYNC_FROM_CC_REF_STORE,
-  SYNC_FROM_CC_CLASS_STORE, SYNC_FROM_CC_CLASS_STORE_AND_REF_STORE, BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD,
+  BROADCAST_TRIGGERED_BY_CC_INSTANCE_SET_GLOBAL_STATE,
+  BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD,
+  SYNC_FROM_CC_CLASS_STORE,
+  SYNC_FROM_CC_CLASS_STORE_AND_REF_STORE,
   SYNC_FROM_GLOBAL_STORE_AND_CC_CLASS_STORE,
   SYNC_FROM_GLOBAL_STORE_AND_CC_CLASS_STORE_AND_REF_STORE,
   SYNC_FROM_GLOBAL_STORE,
@@ -382,13 +385,9 @@ export default function register(ccClassKey, {
               }
             }
 
-            let partialGlobalState = null;
             if (globalStateKeys.length > 0) {
               const { partialState, isStateEmpty: isPartialGlobalStateEmpty } = extractStateByKeys(state, globalStateKeys);
-              if (!isPartialGlobalStateEmpty) {
-                partialGlobalState = partialState;
-                ccContext.store.setGlobalState(partialGlobalState);
-              }
+              if (!isPartialGlobalStateEmpty) ccContext.store.setGlobalState(partialState);
             }
 
             if (this.$$beforeSetState) {
@@ -397,30 +396,30 @@ export default function register(ccClassKey, {
                 // $$beforeSetState(context, next){}
                 this.$$beforeSetState({ changeWay }, () => {
                   this.cc.reactSetState(state, reactCallback);
-                  if (next) next(partialGlobalState);
+                  if (next) next();
                 });
               } else {
                 this.$$beforeSetState({ changeWay });
                 this.cc.reactSetState(state, reactCallback);
-                if (next) next(partialGlobalState);
+                if (next) next();
               }
             } else {
               this.cc.reactSetState(state, reactCallback);
-              if (next) next(partialGlobalState);
+              if (next) next();
             }
           },
-          prepareBroadcastState: (triggerType, moduleName, sourceSharedState, partialGlobalState, needClone) => {
+          prepareBroadcastState: (triggerType, moduleName, sourceSharedState, needClone) => {
             if (this.$$beforeBroadcastState) {
               if (asyncLifeCycleHook) {
-                this.$$beforeBroadcastState({ triggerType, partialGlobalState }, () => {
-                  this.cc.broadcastState(moduleName, sourceSharedState, partialGlobalState, needClone);
+                this.$$beforeBroadcastState({ triggerType }, () => {
+                  this.cc.broadcastState(moduleName, sourceSharedState, needClone);
                 });
               } else {
-                this.$$beforeBroadcastState({ triggerType, partialGlobalState });
-                this.cc.broadcastState(moduleName, sourceSharedState, partialGlobalState, needClone);
+                this.$$beforeBroadcastState({ triggerType });
+                this.cc.broadcastState(moduleName, sourceSharedState, needClone);
               }
             } else {
-              this.cc.broadcastState(moduleName, sourceSharedState, partialGlobalState, needClone);
+              this.cc.broadcastState(moduleName, sourceSharedState, needClone);
             }
           },
           reactSetState: (state, cb) => {
@@ -434,8 +433,10 @@ export default function register(ccClassKey, {
           setState: (state, cb) => {
             this.$$changeState(state, { module: currentModule, cb });
           },
-          setGlobalState: (state, cb) => {
-            this.$$changeState(state, { module: MODULE_GLOBAL, cb });
+          setGlobalState: (partialGlobalState, changeWay = BROADCAST_TRIGGERED_BY_CC_INSTANCE_SET_GLOBAL_STATE) => {
+            // this.cc.prepareBroadcastState(changeWay, module, null, partialGlobalState, false, false);
+            ccContext.store.setGlobalState(partialGlobalState);
+            this.$$changeState(partialGlobalState, { module: currentModule, changeWay });
           },
           forceUpdate: (cb) => {
             this.$$changeState(this.state, { module: currentModule, cb });
@@ -493,7 +494,7 @@ export default function register(ccClassKey, {
               this.cc.invokeWith(reducerFn, { inputModule, forceSync, cb: newCb } = {}, executionContext);
             });
           },
-          broadcastState: (moduleName, sourceSharedState, sourcePartialGlobalState, needClone) => {
+          broadcastState: (moduleName, sourceSharedState, needClone, excludeTriggerRef = true) => {
             let _sourceSharedState = sourceSharedState;
             if (needClone) _sourceSharedState = util.clone(sourceSharedState);// this clone may cause performance issue, if sourceSharedState is too big!!
 
@@ -517,35 +518,38 @@ export default function register(ccClassKey, {
               // extract sourcePartialGlobalState! because different class watch different globalStateKeys
               const {
                 partialState: globalStateForCurrentCcClass, isStateEmpty: isPartialGlobalStateEmpty
-              } = extractStateByKeys(sourcePartialGlobalState, globalStateKeys);
-
-              let toSet = null, changeWay = -1;
+              } = extractStateByKeys(_sourceSharedState, globalStateKeys);
               if (isSharedStateEmpty && isPartialGlobalStateEmpty) return;
+
+              let mergedStateForCurrentCcClass;
               if (isSharedStateEmpty && !isPartialGlobalStateEmpty) {
-                changeWay = SYNC_FROM_CC_INSTANCE_GLOBAL_PARTIAL_STATE;
-                toSet = globalStateForCurrentCcClass;
+                mergedStateForCurrentCcClass = globalStateForCurrentCcClass;
               } else if (!isSharedStateEmpty && isPartialGlobalStateEmpty) {
-                changeWay = SYNC_FROM_CC_INSTANCE_SHARED_STATE;
-                toSet = sharedStateForCurrentCcClass;
-              } else {
-                changeWay = SYNC_FROM_CC_INSTANCE_GLOBAL_PARTIAL_STATE_AND_SHARED_STATE;
-                toSet = { ...globalStateForCurrentCcClass, ...sharedStateForCurrentCcClass };
+                mergedStateForCurrentCcClass = sharedStateForCurrentCcClass;
+              } else {// !isSharedStateEmpty && !isPartialGlobalStateEmpty
+                mergedStateForCurrentCcClass = { ...globalStateForCurrentCcClass, ...sharedStateForCurrentCcClass };
               }
 
-              if (toSet) {
-                ccKeys.forEach(ccKey => {
-                  //exclude currentCcKey, whether its reactSetState has been invoked or not, currentCcKey can't trigger prepareReactSetState here
-                  if (ccKey !== currentCcKey) {
-                    const ref = ccKey_ref_[ccKey];
-                    if (ref) {
-                      const option = ccKey_option_[ccKey];
-                      if (option.syncState) {
-                        ref.cc.prepareReactSetState(changeWay, toSet);
-                      }
-                    }
+              ccKeys.forEach(ccKey => {
+                if (excludeTriggerRef && ccKey === currentCcKey) return;
+                const ref = ccKey_ref_[ccKey];
+                if (ref) {
+                  const option = ccKey_option_[ccKey];
+                  let toSet = null, changeWay = -1;
+                  if (option.syncState && option.syncGlobalState) {
+                    changeWay = SYNC_FROM_CC_INSTANCE_GLOBAL_PARTIAL_STATE_AND_SHARED_STATE;
+                    toSet = mergedStateForCurrentCcClass;
+                  } else if (option.syncState) {
+                    changeWay = SYNC_FROM_CC_INSTANCE_SHARED_STATE;
+                    toSet = sharedStateForCurrentCcClass;
+                  } else if (option.syncGlobalState) {
+                    changeWay = SYNC_FROM_CC_INSTANCE_GLOBAL_PARTIAL_STATE;
+                    toSet = globalStateForCurrentCcClass;
                   }
-                });
-              }
+
+                  if (toSet) ref.cc.prepareReactSetState(changeWay, toSet);
+                }
+              });
 
             });
           }
@@ -576,22 +580,16 @@ export default function register(ccClassKey, {
       //        if ccIns option.syncState is false, cc only change it's own state, 
       //           but if you pass forceSync=true, cc also will broadcast the state to target module and caution: it will overwrite the target module's state !!!
       //        if ccIns option.syncState is true, change it's own state and broadcast the state to target module
-      $$changeState(state, { module, forceSync, cb: reactCallback } = {}) {//executionContext
+      $$changeState(state, { module, changeWay, forceSync, cb: reactCallback } = {}) {//executionContext
         const currentModule = this.cc.ccState.module;
         if (module === currentModule) {
           // who trigger $$changeState, who will go to change the whole received state 
-          this.cc.prepareReactSetState(CHANGE_BY_SELF, state, (partialGlobalState) => {
-            if (this.cc.ccState.ccOption.syncState) {// note!!! if syncState == false, other ccIns will have not change to receive new state
-              this.cc.prepareBroadcastState(BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD, module, state, partialGlobalState, false);
-            } else if (forceSync) {
-              this.cc.prepareBroadcastState(BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD, module, state, partialGlobalState, true);
-            } else {
-              // other ccIns will have no chance to receive new state
-            }
+          this.cc.prepareReactSetState(changeWay || CHANGE_BY_SELF, state, () => {
+            this.cc.prepareBroadcastState(changeWay || BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD, module, state, forceSync);
           }, reactCallback);
         } else {
           if (forceSync) justWarning(`you are trying change another module's state, forceSync=true in not allowed, cc will ignore it!` + vbi(`module:${module} currentModule${currentModule}`));
-          this.cc.prepareBroadcastState(BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD, module, state, partialGlobalState, true);
+          this.cc.prepareBroadcastState(changeWay || BROADCAST_TRIGGERED_BY_CC_INSTANCE_METHOD, module, state, true);
         }
       }